requirements.txt
flask
Pillow
cairosvg
reportlab